native-port: OSRoundUp/Down32B macros + JKRHeap arena size use uintptr_t
OSRoundUp32B/OSRoundDown32B cast their arg (often a pointer) to u32 — truncates under
LP64. Use uintptr_t so pointer alignment is correct on x86-64/arm64. JKRHeap::initArena
computes arena size from two pointers losslessly via uintptr_t. (JKRHeap/JUTException
also contain inline asm → full clean-up lands with the asm-reimplementation pass.)

+++ include/dolphin/os.h
@@ -162,8 +162,8 @@ void OSPanic(const char* file, int line, const char* msg, ...);
 
 #define OSErrorLine(line, ...) OSPanic(__FILE__, line, __VA_ARGS__)
 
-#define OSRoundUp32B(x)   (((u32)(x) + 32 - 1) & ~(32 - 1))
-#define OSRoundDown32B(x) (((u32)(x)) & ~(32 - 1))
+#define OSRoundUp32B(x)   (((uintptr_t)(x) + 32 - 1) & ~(uintptr_t)(32 - 1))
+#define OSRoundDown32B(x) (((uintptr_t)(x)) & ~(uintptr_t)(32 - 1))
 
 void* OSPhysicalToCached(u32 paddr);
 void* OSPhysicalToUncached(u32 paddr);
